Refactored Chart directory structure
