+++ client/src/components/Charts/GamePerDayChart/GamePerDayChart.jsx
@@ -1,4 +1,4 @@
-import { formatDateFull } from '../../utils/helper';
+import { formatDateFull } from '../../../utils/helper';
 import { Line } from 'react-chartjs-2';
 import {
     Chart as ChartJS,
@@ -23,34 +23,34 @@ ChartJS.register(
     Filler
 );
 
-const GuessChart = ({ data }) => {    
-    let guesses = [], dates = [];
+const GamePerDayChart = ({ data }) => {
+    let games = [], dates = [];
     let dataArr = [...data.scoresByUser]
     // sort score array by date
     let sortedData = dataArr.sort((a, b) => { return a.createdAt - b.createdAt })
-    // create array of total guesses per day and array of unique days
+    // create array of total games played per day and array of unique days
     for (let i = 0; i < sortedData.length; i++) {
         if (!dates.includes(formatDateFull(sortedData[i].createdAt))) {
             dates.push(formatDateFull(sortedData[i].createdAt));
-            guesses.push(sortedData[i].guesses);
+            games.push(1);
         } else {
-            const c = guesses.pop();
-            guesses.push(c + sortedData[i].guesses);
+            const c = games.pop();
+            games.push(c + 1);
         };
     };
 
     return (
     <div className='bg-slate-100 rounded-t-md shadow-sm'>
-        <h3 className='bg-slate-300 font-bold p-1 rounded-t-md'>Guesses per Day</h3>
+        <h3 className='bg-slate-300 font-bold p-1 rounded-t-md'>Game per Day</h3>
         <div className='p-1'>
             <Line
                 data={{
                     labels: dates,
                     datasets: [{
-                            label: '# of Guesses',
-                            data: guesses,
-                            borderColor: 'rgb(53, 162, 235)',
-                            backgroundColor: 'rgba(53, 162, 235, 0.5)',
+                            label: '# of Games',
+                            data: games,
+                            borderColor: 'rgb(255, 99, 132)',
+                            backgroundColor: 'rgba(255, 99, 132, 0.5)',
                             fill: true,
                             stepped: true
                         }],
@@ -64,9 +64,7 @@ const GuessChart = ({ data }) => {
                         title: {
                             display: false,
                             text: 'Games Played',
-                            font: {
-                                size: 20,
-                            },
+                            font: { size: 20 },
                             color: '#6b7280',
                         },
                     },
@@ -84,7 +82,7 @@ const GuessChart = ({ data }) => {
                                 maxTicksLimit: 5,
                             },
                             display: false
-                        },
+                        }
                     },
                     color: '#6b7280',
                 }}
@@ -93,4 +91,5 @@ const GuessChart = ({ data }) => {
     </div>
     );
 };
-export default GuessChart;
+
+export default GamePerDayChart;
