+++ client/src/components/Charts/GamePerDayChart/index.js
@@ -1,0 +1,3 @@
+import GamePerDayChart from './GamePerDayChart';
+
+export default GamePerDayChart;
