+++ client/src/components/Charts/GuessPerDayChart/index.js
@@ -1,0 +1,3 @@
+import GuessPerDayChart from './GuessPerDayChart';
+
+export default GuessPerDayChart;
