+++ client/src/components/Charts/index.js
@@ -1,3 +1,5 @@
 import EndGameChart from "./EndGameChart";
+import GamePerDayChart from "./GamePerDayChart";
+import GuessPerDayChart from "./GuessPerDayChart";
 
-export { EndGameChart }
+export { EndGameChart, GamePerDayChart, GuessPerDayChart };
