--- client/src/components/ProfileCharts/index.js
@@ -1,4 +1,0 @@
-import GameChart from "./GameChart";
-import GuessChart from "./GuessChart";
-
-export { GameChart, GuessChart };
